requirements.txt
amazon-textract-textractor
backoff
pymupdf
reportlab
python-dotenv
python-multipart
pydantic
pydantic-settings
fastapi[standard]
boto3-stubs[essentials]
boto3-stubs[s3]
boto3-stubs[textract]